backend/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
psycopg[binary,pool]==3.2.6
werkzeug==2.0.3
langchain-openai
langchain_core
langgraph
langgraph-checkpoint-postgres
langchain
langchain-community
elasticsearch
langchain-elasticsearch
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Cors
google-auth>=2.38.0
google-auth-oauthlib>=1.2.2
google-auth-httplib2>=0.2.0